test changed

+++ spring-database-demo/src/main/java/com/atmik/springdemo/entity/StudentRepository.java
@@ -2,7 +2,7 @@ package com.atmik.springdemo.entity;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface StudentRepository extends CrudRepository<Student, Integer> {
+public interface StudentRepository extends CrudRepository<Student, Long> {
